cleanup, renamed const --> dictionary
