+++ lib/subscribe/collection.js
@@ -1,5 +1,5 @@
 'use strict'
-const REMOVE = require('./const').REMOVE
+const REMOVE = require('./dictionary').REMOVE
 const item = require('./item')
 module.exports = function collection (target, pSubs, subs, update, tree, stamp, rTree, rSubs, force) {
   if (target && target.val !== null) {
