+++ lib/subscribe/dictionary.js
