+++ lib/subscribe/diff.js
@@ -7,5 +7,4 @@ module.exports = function diff (target, subs, update, tree, stamp, rTree, rSubs,
   }
 }
 
-// temp fix (order is convuluted)
 item = require('./item')
