+++ lib/subscribe/item.js
@@ -1,7 +1,5 @@
 'use strict'
 var leaf, collection, struct, root
-// this function can perhaps be optmized -- maybe define the exclusiosn keys
-// what about calling val $val ? -- unclear...
 module.exports = function item (key, target, pSubs, subs, update, tree, stamp, rTree, rSubs, force) {
   if (key !== 'val' && key !== '$' && key !== '_' && key !== 'done') {
     if (key === '$root') {
@@ -16,8 +14,6 @@ module.exports = function item (key, target, pSubs, subs, update, tree, stamp, r
   }
 }
 
-// require order is convoluted now but only way to do it
-// (all call each other etc)
 leaf = require('./leaf')
 collection = require('./collection')
 struct = require('./struct')
