+++ lib/subscribe/leaf.js
@@ -1,10 +1,10 @@
 'use strict'
 const compute = require('./compute')
-const consts = require('./const')
-const INIT = consts.INIT
-const UPDATE = consts.UPDATE
-const REMOVE = consts.REMOVE
-const REMOVEREF = consts.REMOVEREF
+const dictionary = require('./dictionary')
+const INIT = dictionary.INIT
+const UPDATE = dictionary.UPDATE
+const REMOVE = dictionary.REMOVE
+const REMOVEREF = dictionary.REMOVEREF
 
 module.exports = function leaf (key, target, subs, update, tree, stamp, rTree, rSubs, force) {
   var keyTarget = target && target[key]
