+++ lib/subscribe/root/compute.js
@@ -1,6 +1,8 @@
 'use strict'
 const item = require('../item')
-const ROOT = require('../const').ROOT
+const dictionary = require('../dictionary')
+const ROOT = dictionary.ROOT
+const INIT = dictionary.INIT
 
 module.exports = function computeRoot (target, cTree, stamp, rTree, rSubs, istamp, update, force) {
   const rTarget = target.getRoot()
@@ -23,7 +25,7 @@ function execute (target, sTree, stamp, rTree, rSubs, update, force) {
       let typeOverride
       // make tests for remove as well!
       if (!sTree.$r[skey].$) {
-        typeOverride = 'new'
+        typeOverride = INIT
       }
       item(skey, target, subs, subs[skey], function (state, type, stamp, subs, tree) {
         subs.$ = tree.$
