+++ lib/subscribe/struct.js
@@ -1,12 +1,12 @@
 'use strict'
 const compute = require('./compute')
-const consts = require('./const')
-const INIT = consts.INIT
-const UPDATE = consts.UPDATE
-const REMOVE = consts.REMOVE
-const REMOVEREF = consts.REMOVEREF
-const DONE = consts.DONE
 const diff = require('./diff')
+const dictionary = require('./dictionary')
+const INIT = dictionary.INIT
+const UPDATE = dictionary.UPDATE
+const REMOVE = dictionary.REMOVE
+const REMOVEREF = dictionary.REMOVEREF
+const DONE = dictionary.DONE
 
 module.exports = function struct (key, target, pSubs, subs, update, tree, stamp, rTree, rSubs, force) {
   const keyTarget = target && target[key]
@@ -26,12 +26,13 @@ module.exports = function struct (key, target, pSubs, subs, update, tree, stamp,
       treeKey = tree[key] = {}
       leafStamp = compute(keyTarget, treeKey, stamp, rTree, rSubs, update)
       treeKey.$ = leafStamp
-      // make this info a fn
+
       if (ref) {
         treeKey.$ref = traveltaget
       } else if (tree.$ref) {
         treeKey.$ref = tree.$ref[key]
       }
+
       treeKey._p = tree
       treeKey.$ = leafStamp
       if (subs.val) {
@@ -43,6 +44,7 @@ module.exports = function struct (key, target, pSubs, subs, update, tree, stamp,
       }
     } else {
       leafStamp = compute(keyTarget, treeKey, stamp, rTree, rSubs, update)
+
       if (ref) {
         if (
           treeKey.$ref &&
@@ -56,6 +58,7 @@ module.exports = function struct (key, target, pSubs, subs, update, tree, stamp,
       } else if (tree.$ref) {
         treeKey.$ref = tree.$ref[key]
       }
+
       if ((force || treeKey.$) !== leafStamp) {
         treeKey.$ = leafStamp
         if (subs.val === true) {
