+++ test/index.js
@@ -4,4 +4,4 @@ require('./collection')
 require('./references')
 require('./root')
 require('./combined')
-// require('./switch')
+require('./switch')
